+++ oshdb-etl/src/main/java/org/heigit/ohsome/oshdb/tool/importer/transform/oshdb/OSHNode2.java
@@ -95,8 +95,7 @@ public class OSHNode2 extends OSHEntity2 implements OSH<OSMNode> {
         latitude = in.readS64Delta(latitude);
       }
       return new OSMNode(entity.id, version, entity.baseTimestamp + timestamp, changeset, userId,
-          keyValues, (int) (entity.baseLongitude + longitude),
-          (int) (entity.baseLatitude + latitude));
+          keyValues, entity.baseLongitude + longitude, entity.baseLatitude + latitude);
     }
   }
 }
